Merge develop to master

+++ PvPModifier/Network/Packets/PlayerSlotArgs.cs
@@ -6,19 +6,19 @@ using PvPModifier.Variables;
 namespace PvPModifier.Network.Packets {
     public class PlayerSlotArgs : EventArgs {
         public PvPPlayer Player;
-        public int SlotId;
-        public int Stack;
-        public int Prefix;
-        public int NetID;
+        public byte SlotId;
+        public short Stack;
+        public byte Prefix;
+        public short NetID;
 
         public bool ExtractData(MemoryStream data, PvPPlayer player, out PlayerSlotArgs arg) {
             data.ReadByte(); //Passes through the PlayerID data
 
             arg = new PlayerSlotArgs {
-                SlotId = data.ReadByte(),
+                SlotId = (byte)data.ReadByte(),
                 Player = player,
                 Stack = data.ReadInt16(),
-                Prefix = data.ReadByte(),
+                Prefix = (byte)data.ReadByte(),
                 NetID = data.ReadInt16()
             };
